+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [3.5.1] - 2026-08-05
+
+- Added
+  - Bucket, Design Automation engine, and app bundle quick picks now default to your most recently used choice
+  - The "Switch Environment" command now defaults to the last environment used in this workspace
+
 ## [3.5.0] - 2026-08-05
 
 - Added